Added .codecov.yml & new tests
