+++ Tests/FoundationTests/ArrayExtensionsTests.swift
@@ -176,6 +176,9 @@ class ArrayExtensionsTests: XCTestCase {
         let array = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
         let indices = array.indices { $0 % 2 == 0 }
         XCTAssertEqual(indices!, [0, 2, 4, 6, 8])
+		let emptyArray: [Int] = []
+		let emptyIndices = emptyArray.indices { $0 % 2 == 0 }
+		XCTAssertNil(emptyIndices)
     }
     
     func testAllMatch() {
